refactor: cleaner bottombar
